pyproject.toml
[project]
name = "dclite"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "google-genai (>=1.7.0,<2.0.0)",
    "chatterbot (>=1.2.3,<2.0.0)",
    "discord-py-self (>=2.0.1,<3.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "transformers (>=4.50.1,<5.0.0)",
    "huggingface-hub[cli] (>=0.29.3,<0.30.0)",
    "bitsandbytes (>=0.45.4,<0.46.0)",
    "accelerate (>=1.5.2,<2.0.0)",
    "notebook (>=7.3.3,<8.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "wordcloud (>=1.9.4,<2.0.0)"
]

[tool.poetry]
packages = [{include = "dclite", from = "src"}]


[tool.poetry.group.dev.dependencies]
dotenv = "^0.9.9"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
